Added Event Deletion
Events can be deleted.
To delete an event, click on it in the calendar to see more details about the event.
The event details pop up menu shows a delete button that when clicked shows a confirm message to delete the event. If it is confirmed the event is then removed from the calendar.