Fixed Event def and Event schedule bugs
I disabled the delete button from event definitions that have been used to grant competencies to employees as this violates a foreign key constraint.

I also fixed the delete bug on the Calendar page that originally did not work for deleting an event due to the multi day styling logically error.